o Updated scoreboard weapon icons o Added cl_iconr, cl_icong, cl_iconb to control scoreboard icon colour for testing ( to be removed before release ) o Fixed compile issue with textmessage change.
git-svn-id: https://unknownworlds.svn.cloudforge.com/ns1@394 67975925-1194-0748-b3d5-c16f83f1a3a1

+++ releases/3.2.0/source/cl_dll/text_message.cpp
@@ -178,7 +178,7 @@ int CHudTextMessage::MsgFunc_TextMsg( const char *pszName, int iSize, void *pbuf
 	}
 
 	// Ensure that message[0] does not contain exessive %'s, max 4x%s, all other %'s removed.
-	int lastpos = 0; int pos; int count = 0;
+	size_t lastpos = 0; size_t pos; int count = 0;
 	while(true)
 	{
 		pos = message[0].find("%", lastpos);
